modularize forgot password and change password into separate components
