+++ src/App.tsx
@@ -1,9 +1,11 @@
 import { Routes, Route } from 'react-router-dom'
 import { Layout } from 'layout'
 import {
+  ChangePassword,
   EmailSentConfirmation,
   EmailVerify,
   ExpiredLink,
+  ResetPassword,
   SignIn,
   SignUp,
 } from 'screens/Auth'
@@ -19,11 +21,9 @@ import { SnackBarToast } from 'components/SnackBarToast/SnackBarToast'
 import { UnassignedProject } from 'screens/Project/UnassignedProject'
 import { AboutUs, HowTo } from 'screens/Landing'
 import { Email } from 'screens/Auth/Settings/Email'
-import { PasswordSettings } from 'screens/Auth/Settings/Password'
 import { Account } from 'screens/Auth/Settings/Account'
 import { TeamMembers } from 'screens/Project/TeamMembers'
 import { ProjectDetails } from 'screens/Project/ProjectDetails'
-import { ResetPassword } from 'screens/Auth/Settings/Password'
 import { SuccessScreen } from 'screens/SuccessScreen/SuccessScreen'
 
 function App() {
@@ -58,7 +58,7 @@ function App() {
           />
           <Route path='/users/:id/settings'>
             <Route path='email' element={<Email />} />
-            <Route path='password' element={<PasswordSettings />} />
+            <Route path='password' element={<ChangePassword />} />
             <Route path='account' element={<Account />} />
           </Route>
           {/* Project */}
