+++ src/screens/Auth/Settings/ChangePassword.tsx
@@ -1,0 +1,107 @@
+import './Settings.scss'
+import { PasswordFormData } from 'interfaces/AccountSettingsInterface'
+import { PasswordErrors } from 'interfaces/components'
+import { useEffect, useState } from 'react'
+import { useDispatch } from 'react-redux'
+import { logOut, updateUsersPassword } from 'utils/api'
+import { SuccessQueryParam } from 'utils/data/authSettingsConstants'
+import { emptyPasswordData } from 'utils/data/userConstants'
+import { useAppSelector } from 'utils/redux/hooks'
+import { createSnackBar } from 'utils/redux/slices/snackBarSlice'
+import { logoutAuthUser, selectAuthUser } from 'utils/redux/slices/userSlice'
+import { useFormValidation } from 'utils/helpers'
+import { PasswordInputs } from 'components/Inputs'
+import { ThemeProvider } from '@emotion/react'
+import { Button, createTheme } from '@mui/material'
+import { useNavigate } from 'react-router-dom'
+
+export const ChangePassword = () => {
+  const navigate = useNavigate()
+  const [formValues, setFormValues] =
+    useState<PasswordFormData>(emptyPasswordData)
+  const [passwordErrors, setPasswordErrors] = useState<PasswordErrors>({})
+  const [inputError, setInputError] = useState<boolean>(false)
+  const [isDisabled, toggleIsDisabled] = useState(false)
+  const authUser = useAppSelector(selectAuthUser)
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    setFormValues(emptyPasswordData)
+  }, [])
+
+  const handleSubmit = async e => {
+    e.preventDefault()
+
+    const reqBody = {
+      password: formValues.currentPassword,
+      newPassword: formValues.password,
+      confirmNewPassword: formValues.confirmPassword,
+    }
+    const passwordData = await updateUsersPassword(reqBody, authUser._id)
+
+    if (passwordData.status >= 400) {
+      passwordData.friendlyMessage === 'Your password is incorrect.' &&
+        setInputError(true)
+
+      dispatch(
+        createSnackBar({
+          isOpen: true,
+          message: passwordData.friendlyMessage,
+          duration: 5000,
+          severity: 'error',
+          horizontal: 'right',
+        })
+      )
+    } else {
+      await logOut()
+      dispatch(logoutAuthUser())
+      setInputError(false)
+      navigate(
+        `/success/${authUser._id}?screen=${SuccessQueryParam.changePassword}`
+      )
+    }
+  }
+
+  const resetErrorState = () => {
+    setInputError(false)
+  }
+
+  const { password, currentPassword } = formValues
+
+  const theme = createTheme({
+    palette: {
+      primary: {
+        main: '#FFA726',
+      },
+    },
+  })
+
+  useFormValidation(formValues, currentPassword, toggleIsDisabled)
+  return (
+    <div className='settings-change-password container'>
+      <form className='settings-change-password form' onSubmit={handleSubmit}>
+        <div className='settings-change-password header'>Change password</div>
+        <PasswordInputs
+          disableErrorState={resetErrorState}
+          formValues={formValues}
+          inputError={inputError}
+          password={password}
+          passwordErrors={passwordErrors}
+          setPasswordErrors={setPasswordErrors}
+          setFormValues={setFormValues}
+          passwordInputName='settings-pwd-reset'
+        />
+        <ThemeProvider theme={theme}>
+          <Button
+            className='settings-change-password button'
+            variant='contained'
+            type='submit'
+            disabled={isDisabled}
+          >
+            Change password
+          </Button>
+        </ThemeProvider>
+      </form>
+    </div>
+  )
+}
