+++ src/screens/Auth/Settings/index.ts
@@ -1,3 +1,4 @@
 export * from './Account'
 export * from './Email'
-export * from './Password'
+export * from './ChangePassword'
+export * from './ResetPassword'
