+++ src/utils/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './calendarHelpers'
 export * from './stateHelpers'
 export * from './routeHelpers'
 export * from './projectHelpers'
+export * from './passwordHelpers'
